pyproject.toml
[project]
name = "ubik"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "kronicler>=0.1.3",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "ruff>=0.14.11",
    "hy>=1.2.0",
    "requests>=2.32.5",
]
